feat: Integrate SMS Opt Out feature into phone number handling in Create and Edit student forms

+++ app-modules/student-data-model/src/Filament/Resources/StudentResource/Pages/EditStudent.php
@@ -200,44 +200,52 @@ class EditStudent extends EditRecord
                                     ->visible(SmsOptOutFeature::active()),
                             ])
                             ->mutateRelationshipDataBeforeFillUsing(function (array $data): array {
-                                if (isset($data['number'])) {
-                                    $data['sms_opt_out_phone_number'] = SmsOptOutPhoneNumber::where('number', $data['number'])->exists();
-                                } else {
-                                    $data['sms_opt_out_phone_number'] = false;
+                                if (SmsOptOutFeature::active()) {
+                                    if (isset($data['number'])) {
+                                        $data['sms_opt_out_phone_number'] = SmsOptOutPhoneNumber::where('number', $data['number'])->exists();
+                                    } else {
+                                        $data['sms_opt_out_phone_number'] = false;
+                                    }
                                 }
 
                                 return $data;
                             })
                             ->mutateRelationshipDataBeforeCreateUsing(function (array $data): array {
-                                if ($data['sms_opt_out_phone_number'] === true) {
-                                    SmsOptOutPhoneNumber::firstOrCreate([
-                                        'number' => $data['number'],
-                                    ]);
-                                }
+                                if (SmsOptOutFeature::active()) {
+                                    if ($data['sms_opt_out_phone_number'] === true) {
+                                        SmsOptOutPhoneNumber::firstOrCreate([
+                                            'number' => $data['number'],
+                                        ]);
+                                    }
 
-                                unset($data['sms_opt_out_phone_number']);
+                                    unset($data['sms_opt_out_phone_number']);
+                                }
 
                                 return $data;
                             })
                             ->mutateRelationshipDataBeforeSaveUsing(function (array $data): array {
-                                if ($data['sms_opt_out_phone_number'] === true) {
-                                    SmsOptOutPhoneNumber::firstOrCreate([
-                                        'number' => $data['number'],
-                                    ]);
-                                } else {
-                                    SmsOptOutPhoneNumber::where('number', $data['number'])->delete();
-                                }
+                                if (SmsOptOutFeature::active()) {
+                                    if ($data['sms_opt_out_phone_number'] === true) {
+                                        SmsOptOutPhoneNumber::firstOrCreate([
+                                            'number' => $data['number'],
+                                        ]);
+                                    } else {
+                                        SmsOptOutPhoneNumber::where('number', $data['number'])->delete();
+                                    }
 
-                                unset($data['sms_opt_out_phone_number']);
+                                    unset($data['sms_opt_out_phone_number']);
+                                }
 
                                 return $data;
                             })
                             ->deleteAction(
                                 fn (Action $action) => $action->before(function () use ($action) {
-                                    $data = $action->getArguments();
+                                    if (SmsOptOutFeature::active()) {
+                                        $data = $action->getArguments();
 
-                                    if (isset($data['number'])) {
-                                        SmsOptOutPhoneNumber::where('number', $data['number'])->delete();
+                                        if (isset($data['number'])) {
+                                            SmsOptOutPhoneNumber::where('number', $data['number'])->delete();
+                                        }
                                     }
                                 })
                             )
